agent_factory/requirements.txt
python-dotenv==1.0.1
rich==13.7.1
requests==2.32.3
pandas==2.2.2
Pillow==10.4.0
pydub==0.25.1
yt-dlp==2024.4.9
google-api-python-client==2.123.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
huggingface-hub==0.23.4
langchain-core==0.2.28
langchain-community==0.2.16
moviepy==1.0.3
ffmpeg-python==0.2.0
python-multipart==0.0.9
